Israeli baby names per year.

Full baby name data provided by the Israel's Central Bureau of Statistics (LAMAS). Only names with at least 5 uses (in a given year) are included. Data is separated by sectors and sex, where sectors are "Jewish", "Muslim", "Christian", "Druze" and "Other".

Usage

babynamesIL

Format

A data frame with six variables: sector, year, sex, name, n and prop (n divided by total number of babies in the database).

Details

Data was downloaded from: here

2022 data was downloaded from: here

2023 data was downloaded from: here